Infra quickfix (#123)
* Returning max 100HP
(accounting for >100HP before hp decay)

* camelCaseFFS

* Revert "camelCaseFFS"

This reverts commit 83705685ebf68e1e7749866fa078078eccf9b753.

* camelCasePls

* camelCase p2

* Rename utilfunctions.go to utilFunctions.go

* Rename pkg/utils/utilfunctions/utilFunctions.go to pkg/utils/utilFunctions/utilFunctions.go

* more general min int func
