+++ pkg/infra/baseagent.go
@@ -10,6 +10,7 @@ import (
 	"github.com/SOMAS2021/SOMAS2021/pkg/utils/globalTypes/food"
 	"github.com/SOMAS2021/SOMAS2021/pkg/utils/globalTypes/health"
 	"github.com/SOMAS2021/SOMAS2021/pkg/utils/globalTypes/world"
+	"github.com/SOMAS2021/SOMAS2021/pkg/utils/utilFunctions"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -81,7 +82,7 @@ func (a *Base) Run() {
 }
 
 func (a *Base) HP() int {
-	return a.hp
+	return utilFunctions.MinInt(a.hp, 100)
 }
 
 // only show the food on the platform if the platform is on the
