+++ pkg/utils/utilFunctions/utilFunctions.go
@@ -1,0 +1,20 @@
+package utilFunctions
+
+func Sum(inputList []int) int {
+	totalAgents := 0
+	for _, value := range inputList {
+		totalAgents += value
+	}
+	return totalAgents
+
+}
+
+func MinInt(vars ...int) int {
+	min := vars[0]
+	for _, i := range vars {
+		if min > i {
+			min = i
+		}
+	}
+	return min
+}
