--- pkg/utils/utilfunctions/utilfunctions.go
@@ -1,10 +1,0 @@
-package utilfunctions
-
-func Sum(inputList []int) int {
-	totalAgents := 0
-	for _, value := range inputList {
-		totalAgents += value
-	}
-	return totalAgents
-
-}
